Made few changes for Customer Wallet Assignment.
